Changing k8s secret OpenStack-Exporter mounts
The normal keystone secret that is mounted into the OpenStack
exporter uses the default admin username for monitoring OS instances.
This changes the secret to use novas - which contains the credentials
for our charter_admin user we have bootstrapped.  This allows us to
have visibility into the project where our instances are being created.